% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oVisEvidenceAdv.r
\name{oVisEvidenceAdv}
\alias{oVisEvidenceAdv}
\title{Function to visualise evidence and priority scores for prioritised genes in a gene network}
\usage{
oVisEvidenceAdv(
xTarget,
g = NA,
nodes = NULL,
node.info = c("smart", "simple", "none"),
neighbor.order = 1,
neighbor.seed = TRUE,
neighbor.top = NULL,
largest.comp = TRUE,
node.label.size = 2,
node.label.color = "steelblue4",
node.label.alpha = 0.9,
node.label.padding = 0.5,
node.label.arrow = 0,
node.label.force = 0.1,
node.shape = 19,
node.color.title = "Pi\\nrating",
colormap = "white-yellow-red",
ncolors = 64,
zlim = c(0, 5),
node.size.range = 5,
title = "",
edge.color = "steelblue",
edge.color.alpha = 0.2,
edge.curve = 0,
edge.arrow.gap = 0.025,
pie.radius = NULL,
pie.color = "black",
pie.color.alpha = 1,
pie.thick = 0.1,
...
)
}
\arguments{
\item{xTarget}{an object of class "dTarget", "sTarget" or "eTarget"}

\item{g}{an object of class "igraph". If NA, the 'metag' will be used,
which is part of the input object "xTarget". If provided, it must have
node attributes ('priority','xcoord','ycoord')}

\item{nodes}{which node genes are in query. If NULL, the top gene will
be queried}

\item{node.info}{tells the additional information used to label nodes.
It can be one of "none" (only gene labeling), "smart" for (by default)
using three pieces of information (if any): genes, 5-star ratings, and
associated ranks (marked by an @ icon)}

\item{neighbor.order}{an integer giving the order of the neighborhood.
By default, it is 1-order neighborhood}

\item{neighbor.seed}{logical to indicate whether neighbors are seeds
only. By default, it sets to true}

\item{neighbor.top}{the top number of the neighbors with the highest
priority. By default, it sets to NULL to disable this parameter}

\item{largest.comp}{logical to indicate whether the largest component
is only retained. By default, it sets to true for the largest component
being left}

\item{node.label.size}{a vector specifying node size or a character
specifying which node attribute used for node label size}

\item{node.label.color}{the node label color}

\item{node.label.alpha}{the 0-1 value specifying transparency of node
labelling}

\item{node.label.padding}{the padding around the labeled node}

\item{node.label.arrow}{the arrow pointing to the labeled node}

\item{node.label.force}{the repelling force between overlapping labels}

\item{node.shape}{an integer specifying node shape}

\item{node.color.title}{a character specifying the title for node
coloring}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap),
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta), and "ggplot2" (emulating ggplot2
default color palette). Alternatively, any hyphen-separated HTML color
names, e.g. "lightyellow-orange" (by default), "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{https://html-color-codes.info/color-names/index.html}}

\item{ncolors}{the number of colors specified over the colormap}

\item{zlim}{the minimum and maximum values for which colors should be
plotted}

\item{node.size.range}{the range of actual node size}

\item{title}{a character specifying the title for the plot}

\item{edge.color}{a character specifying the edge colors}

\item{edge.color.alpha}{the 0-1 value specifying transparency of edge
colors}

\item{edge.curve}{a numeric value specifying the edge curve. 0 for the
straight line}

\item{edge.arrow.gap}{a gap between the arrow and the node}

\item{pie.radius}{the radius of a pie. If NULL, it equals roughly 1/75}

\item{pie.color}{the border color of a pie}

\item{pie.color.alpha}{the 0-1 value specifying transparency of pie
border colors}

\item{pie.thick}{the pie border thickness}

\item{...}{additional graphic parameters for oGGnetwork}
}
\value{
a ggplot object.
}
\description{
\code{oVisEvidenceAdv} is supposed to visualise evidence and priority
scores for prioritised genes in a gene network. It returns an object of
class "ggplot".
}
\examples{
\dontrun{
## TNFRSF1A
oVisEvidenceAdv(xTarget, nodes="TNFRSF1A", neighbor.order=1,
neighbor.seed=TRUE, neighbor.top=NULL)
}
}
\seealso{
\code{\link{oVisEvidence}}
}
